fix: page-fault caused by memory leak in `pfree()`
Because of the memory leaking in `pfree()`, it forget to reset the memory bitmap flag after free a page.
That cause every time when the program yield to `idle` thread and then switch back and then reload the page
dir into register `CR3`, the page next to the previous page that we forgot the free will be clear, set to 0.
That sequently cause the information of arena in that page lost. Then when we free that block of memory, it
will cause page fault because of descriptor in arena is set to NULL(0).

But actually, I don't know why reload the page dir address into `CR3` register will cause that. I plan to dig in
it.